requirements.txt
# Core app
streamlit==1.37.1
requests>=2.31.0,<3
python-dotenv>=1.0.1,<2

# Optional smoother countdown refresh (used in app; falls back to JS if absent)
streamlit-autorefresh>=1.0.1

# LLM client (you call OpenAI via requests, but keep this if you switch to the SDK)
openai==1.43.0

# RAG (optional but supported)
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
pypdf==4.3.1
numpy==1.26.4
